FTX Positions clean up (there was 2 FtxPositionsDto which clashed)
