+++ xchange-ftx/src/main/java/org/knowm/xchange/ftx/dto/account/FtxPositionDto.java
@@ -1,6 +1,8 @@
 package org.knowm.xchange.ftx.dto.account;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.knowm.xchange.ftx.dto.trade.FtxOrderSide;
+
 import java.math.BigDecimal;
 
 public class FtxPositionDto {
@@ -10,6 +12,9 @@ public class FtxPositionDto {
 
   @JsonProperty("entryPrice")
   private final BigDecimal entryPrice;
+
+  @JsonProperty("estimatedLiquidationPrice")
+  private final BigDecimal estimatedLiquidationPrice;
 
   @JsonProperty("future")
   private final String future;
@@ -36,7 +41,7 @@ public class FtxPositionDto {
   private final BigDecimal shortOrderSize;
 
   @JsonProperty("side")
-  private final String side;
+  private final FtxOrderSide side;
 
   @JsonProperty("size")
   private final BigDecimal size;
@@ -44,9 +49,13 @@ public class FtxPositionDto {
   @JsonProperty("unrealizedPnl")
   private final BigDecimal unrealizedPnl;
 
+  @JsonProperty("collateralUsed")
+  private final BigDecimal collateralUsed;
+
   public FtxPositionDto(
       @JsonProperty("cost") BigDecimal cost,
       @JsonProperty("entryPrice") BigDecimal entryPrice,
+      @JsonProperty("estimatedLiquidationPrice") BigDecimal estimatedLiquidationPrice,
       @JsonProperty("future") String future,
       @JsonProperty("initialMarginRequirement") BigDecimal initialMarginRequirement,
       @JsonProperty("longOrderSize") BigDecimal longOrderSize,
@@ -55,11 +64,13 @@ public class FtxPositionDto {
       @JsonProperty("openSize") BigDecimal openSize,
       @JsonProperty("realizedPnl") BigDecimal realizedPnl,
       @JsonProperty("shortOrderSize") BigDecimal shortOrderSize,
-      @JsonProperty("side") String side,
+      @JsonProperty("side") FtxOrderSide side,
       @JsonProperty("size") BigDecimal size,
-      @JsonProperty("unrealizedPnl") BigDecimal unrealizedPnl) {
+      @JsonProperty("unrealizedPnl") BigDecimal unrealizedPnl,
+      @JsonProperty("collateralUsed") BigDecimal collateralUsed) {
     this.cost = cost;
     this.entryPrice = entryPrice;
+    this.estimatedLiquidationPrice = estimatedLiquidationPrice;
     this.future = future;
     this.initialMarginRequirement = initialMarginRequirement;
     this.longOrderSize = longOrderSize;
@@ -71,6 +82,7 @@ public class FtxPositionDto {
     this.side = side;
     this.size = size;
     this.unrealizedPnl = unrealizedPnl;
+    this.collateralUsed = collateralUsed;
   }
 
   public BigDecimal getCost() {
@@ -79,6 +91,10 @@ public class FtxPositionDto {
 
   public BigDecimal getEntryPrice() {
     return entryPrice;
+  }
+
+  public BigDecimal getEstimatedLiquidationPrice() {
+    return estimatedLiquidationPrice;
   }
 
   public String getFuture() {
@@ -113,7 +129,7 @@ public class FtxPositionDto {
     return shortOrderSize;
   }
 
-  public String getSide() {
+  public FtxOrderSide getSide() {
     return side;
   }
 
@@ -125,6 +141,10 @@ public class FtxPositionDto {
     return unrealizedPnl;
   }
 
+  public BigDecimal getCollateralUsed() {
+    return collateralUsed;
+  }
+
   @Override
   public String toString() {
     return "FtxPositionDto{"
@@ -132,9 +152,10 @@ public class FtxPositionDto {
         + cost
         + ", entryPrice="
         + entryPrice
-        + ", future='"
+        + ", estimatedLiquidationPrice="
+        + estimatedLiquidationPrice
+        + ", future="
         + future
-        + '\''
         + ", initialMarginRequirement="
         + initialMarginRequirement
         + ", longOrderSize="
@@ -149,13 +170,14 @@ public class FtxPositionDto {
         + realizedPnl
         + ", shortOrderSize="
         + shortOrderSize
-        + ", side='"
+        + ", side="
         + side
-        + '\''
         + ", size="
         + size
         + ", unrealizedPnl="
         + unrealizedPnl
+        + ", collateralUsed="
+        + collateralUsed
         + '}';
   }
 }
